+++ cypress/e2e/select-one.spec.ts
@@ -1070,25 +1070,6 @@ describe('Choices - select one', () => {
     });
 
     describe('allow html', () => {
-      describe('is undefined', () => {
-        it('logs a deprecation warning', () => {
-          cy.get('@consoleWarn').should(
-            'be.calledOnceWithExactly',
-            'Deprecation warning: allowHTML will default to false in a future release. To render HTML in Choices, you will need to set it to true. Setting allowHTML will suppress this message.',
-          );
-        });
-
-        it('does not show html as text', () => {
-          cy.get('[data-test-hook=allowhtml-undefined]')
-            .find('.choices__list--dropdown .choices__list')
-            .children()
-            .first()
-            .should(($choice) => {
-              expect($choice.text().trim()).to.equal('Choice 1');
-            });
-        });
-      });
-
       describe('set to true', () => {
         it('does not show html as text', () => {
           cy.get('[data-test-hook=allowhtml-true]')