+++ cypress/e2e/text.spec.ts
@@ -360,24 +360,6 @@ describe('Choices - text element', () => {
     });
 
     describe('allow html', () => {
-      describe('is undefined', () => {
-        it('logs a deprecation warning', () => {
-          cy.get('@consoleWarn').should(
-            'be.calledOnceWithExactly',
-            'Deprecation warning: allowHTML will default to false in a future release. To render HTML in Choices, you will need to set it to true. Setting allowHTML will suppress this message.',
-          );
-        });
-
-        it('does not show html as text', () => {
-          cy.get('[data-test-hook=allowhtml-undefined]')
-            .find('.choices__list--multiple .choices__item')
-            .first()
-            .should(($choice) => {
-              expect($choice.text().trim()).to.equal('Mason Rogers');
-            });
-        });
-      });
-
       describe('set to true', () => {
         it('does not show html as text', () => {
           cy.get('[data-test-hook=allowhtml-true]')